terra-auth-service/requirements.txt
asgiref==3.10.0
bcrypt==3.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==2.0.12
Django==3.2.9
djangorestframework==3.12.4
idna==3.11
mysqlclient==2.1.1
pycparser==2.23
PyJWT==2.3.0
python-dotenv==0.19.0
pytz==2025.2
requests==2.26.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==1.26.20
